Create AmountDistribution.apxt
Trigger Handler :

How to create a new trigger :

While still in the trailhead account, navigate to the gear icon in the top right corner.

Click on developer console and you will be navigated to a new console window.

Click on File menu in the tool bar, and click on new? Trigger.

Enter the trigger name and the object to be triggered.

Name  : AmountDistribution

sObject : Appointment__c


Syntax For creating trigger :


The syntax for creating trigger is :


Trigger [trigger name] on [object name]( Before/After event)

{


}

In this project , trigger is called whenever the particular records sum exceed the threshold i.e minimum business requirement value. Then the code in the trigger will get executed.